Remove support for 32bit
x86-32 is rapidly losing support, with Ubuntu recently announcing they
won't be building 32bit distros or compatibility libraries. While it was
a nice sanity check to cross-compile to 32bit it does create a burden of
having to maintain two different in-memory layouts for our core runtime
types. It also requires a significant amount of work on CI.

This kills 32bit support by:

1. Only have a single memory layout for all of our runtime times

2. Removing the MIR "usize" type and just using `Int64` everywhere

3. Removing our i686 Buildkite script from CI

+++ runtime/intern.rs
@@ -28,14 +28,15 @@ const INLINE_SIZE: usize = 8;
 
 #[repr(C)]
 pub struct GlobalName {
-    name_byte_length: usize,
+    name_byte_length: u64,
     name_bytes: *const u8,
 }
 
 impl GlobalName {
     fn as_str(&self) -> &str {
         unsafe {
-            let byte_slice = std::slice::from_raw_parts(self.name_bytes, self.name_byte_length);
+            let byte_slice =
+                std::slice::from_raw_parts(self.name_bytes, self.name_byte_length as usize);
             std::str::from_utf8_unchecked(byte_slice)
         }
     }
